backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
transformers
pillow
scikit-learn
numpy
pandas